sedgewick/src/test/java/sedgewick/basic/problems/stack/PostFixEvalTest.java
package sedgewick.basic.problems.stack;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class PostFixEvalTest {

    @Test
    void testBasic() throws NoSuchFieldException {
        final String postFix = "2 3 +";
        Assertions.assertEquals(5, PostFixEval.eval(postFix));
    }

    @Test
    void testSimple() throws NoSuchFieldException {
        final String postFix = "3 + 4 * 5 / 6";
        Assertions.assertEquals(6.333333333, PostFixEval.eval(postFix), 0.0001);
    }

    @Test
    void testMedium() throws NoSuchFieldException {
        final String postFix = "(300 + 23) * (43 - 21)/(84 + 7)";
        Assertions.assertEquals(78.0879, PostFixEval.eval(postFix), 0.0001);
    }

    @Test
    void testComplex() throws NoSuchFieldException {
        final String postFix = "(4 + 8) * (6 - 5)/((3 - 2)*(2 + 2))";
        Assertions.assertEquals(3, PostFixEval.eval(postFix), 0.0001);
    }
}